Fix bugs with maintaining refs
Since I wasn't aware that when components are unmounted, their ref
callbacks are invoked with `null`, everything was out of whack. With
that knowledge, it turns out that a much cleaner data structure for
`this._tags` in the master `TagList` is an object that maps from Tag key
to jQuery wrapped ref.

Also remove the README in the app.